refactor: notification 카드 상수/유틸 분리
