+++ apps/web/src/features/notification/index.ts
@@ -1,3 +1,4 @@
 export { NotificationCard } from './ui/NotificationCard';
 export { NotificationCardList } from './ui/NotificationCardList';
 export { NotificationBell } from './ui/NotificationBell';
+export { CATEGORY_LABEL } from './model/categoryLabel';
