+++ apps/web/src/features/notification/model/categoryLabel.ts
@@ -1,0 +1,7 @@
+import type { NotificationType } from '@/entities/notification';
+
+export const CATEGORY_LABEL: Record<NotificationType, string> = {
+  RETROSPECT: '소비 돌아보기',
+  WEEKLY_REPORT: '분석 리포트 보러가기',
+  SERVICE_NOTICE: '',
+};
