+++ apps/web/src/features/notification/ui/NotificationCard/NotificationCard.tsx
@@ -3,21 +3,9 @@
 import React from 'react';
 import { NotificationCard as NotificationCardUI } from '@/shared/ui/notificationCard';
 import { IcAlertReport, IcNotice } from 'public/icons';
-import type { NotificationResponseDTO, NotificationType } from '@/entities/notification';
-
-const CATEGORY_LABEL: Record<NotificationType, string> = {
-  RETROSPECT: '소비 돌아보기',
-  WEEKLY_REPORT: '분석 리포트 보러가기',
-  SERVICE_NOTICE: '',
-};
-
-const formatDate = (dateStr?: string) => {
-  if (!dateStr) return '';
-  const date = new Date(dateStr);
-  const month = String(date.getMonth() + 1).padStart(2, '0');
-  const day = String(date.getDate()).padStart(2, '0');
-  return `${month}월 ${day}일`;
-};
+import type { NotificationResponseDTO } from '@/entities/notification';
+import { formatDateToMonthDay } from '@/shared/utils';
+import { CATEGORY_LABEL } from '../../model/categoryLabel';
 
 interface NotificationCardProps {
   notification: NotificationResponseDTO;
@@ -38,7 +26,7 @@ export const NotificationCard = ({ notification }: NotificationCardProps) => {
         icon={icon}
         title={title ?? ''}
         message={message ?? ''}
-        date={formatDate(createdAt)}
+        date={formatDateToMonthDay(createdAt)}
         actionLabel={actionLabel}
         url={redirectUrl}
       />
