+++ apps/web/src/shared/utils/date.ts
@@ -18,3 +18,14 @@ export const formatDate = (date: Date): string => {
   const day = date.getDate();
   return `${year}년 ${month}월 ${day}일`;
 };
+
+/**
+ * ISO 날짜 문자열을 "MM월 DD일" 형식의 문자열로 변환
+ */
+export const formatDateToMonthDay = (dateStr?: string): string => {
+  if (!dateStr) return '';
+  const date = new Date(dateStr);
+  const month = String(date.getMonth() + 1).padStart(2, '0');
+  const day = String(date.getDate()).padStart(2, '0');
+  return `${month}월 ${day}일`;
+};
