+++ apps/web/src/shared/utils/index.ts
@@ -1,5 +1,5 @@
 export { getTextWidth } from './getTextWidth';
 export { normalizeNumberValue, formatNumberWithComma } from './inputFormatters';
-export { formatDate, formatDateToISO } from './date';
+export { formatDate, formatDateToISO, formatDateToMonthDay } from './date';
 export { getPlatform } from './getPlatform';
 export { clearAllUserStorage, USER_STORAGE_KEYS } from './clearUserStorage';
